fix: Address remaining CodeRabbitAI review feedback
Critical fixes:
- Fixed race condition in runInParallel using Set for proper promise cleanup
- Removed fixed sleeps in logout flow, now uses event-driven waits
- Extracted duplicate registerAndLogin helper to shared auth-helpers.ts
- Removed unused testUser variable in notes.spec.ts

All quality checks passing:
✅ TypeScript compilation
✅ Linting (0 errors, 0 warnings)
✅ Build successful
✅ Unit tests (267 tests passing)
